--- bundles/org.smarthomej.binding.basicprofiles/src/main/java/org/smarthomej/binding/basicprofiles/internal/config/RoundStateProfileConfig.java
@@ -1,30 +1,0 @@
-/**
- * Copyright (c) 2010-2021 Contributors to the openHAB project
- *
- * See the NOTICE file(s) distributed with this work for additional
- * information.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0
- *
- * SPDX-License-Identifier: EPL-2.0
- */
-package org.smarthomej.binding.basicprofiles.internal.config;
-
-import java.math.RoundingMode;
-
-import org.eclipse.jdt.annotation.NonNullByDefault;
-import org.eclipse.jdt.annotation.Nullable;
-import org.smarthomej.binding.basicprofiles.internal.profiles.RoundStateProfile;
-
-/**
- * Configuration for {@link RoundStateProfile}.
- *
- * @author Christoph Weitkamp - Initial contribution
- */
-@NonNullByDefault
-public class RoundStateProfileConfig {
-    public @Nullable Integer scale;
-    public String mode = RoundingMode.HALF_UP.name();
-}
