+++ bundles/org.smarthomej.transform.basicprofiles/NOTICE
