+++ bundles/org.smarthomej.transform.basicprofiles/pom.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0"
+  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+
+  <modelVersion>4.0.0</modelVersion>
+
+  <parent>
+    <groupId>org.smarthomej.addons.bundles</groupId>
+    <artifactId>org.smarthomej.addons.reactor.bundles</artifactId>
+    <version>3.2.0-SNAPSHOT</version>
+  </parent>
+
+  <artifactId>org.smarthomej.transform.basicprofiles</artifactId>
+
+  <name>SmartHome/J Add-ons :: Bundles :: Transformation Service :: Basic Profiles</name>
+
+</project>
