+++ bundles/org.smarthomej.transform.basicprofiles/src/main/feature/feature.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<features name="org.smarthomej.transform.basicprofiles-${project.version}" xmlns="http://karaf.apache.org/xmlns/features/v1.4.0">
+	<repository>mvn:org.openhab.core.features.karaf/org.openhab.core.features.karaf.openhab-core/${ohc.version}/xml/features</repository>
+
+	<feature name="smarthomej-transformation-basicprofiles" description="Basic Profiles" version="${project.version}">
+		<feature>openhab-runtime-base</feature>
+		<bundle start-level="80">mvn:org.smarthomej.addons.bundles/org.smarthomej.transform.basicprofiles/${project.version}</bundle>
+	</feature>
+</features>
