+++ bundles/org.smarthomej.transform.basicprofiles/src/main/resources/OH-INF/binding/binding.xml
