+++ bundles/org.smarthomej.transform.basicprofiles/src/main/resources/OH-INF/config/config.xml
