+++ bundles/org.smarthomej.transform.basicprofiles/src/main/resources/OH-INF/config/timestampOffsetProfile.xml
